requirements.txt
streamlit
pandas
edinet-python
numpy
beautifulsoup4>=4.8.2
joblib>=0.14.1
lxml>=4.4.2
requests>=2.22.0
tqdm>=4.41.1